demo(journalist): refactor into modules
